se probó el correcto funcionamiento de lo que se programó en los metodos store y update del controlador ProductsController

+++ app/Http/Controllers/Dashboard/ProductsController.php
@@ -9,7 +9,7 @@ use Vest\Http\Controllers\Controller;
 
 //tablas a usar
 use Vest\Tables\Product;
-use Vest\Tables\User;
+use Vest\User;
 
 //Request para validar
 use Vest\Http\Requests\CreateProductRequest;
@@ -89,32 +89,39 @@ class ProductsController extends Controller
      */
     public function store(CreateProductRequest $request)
     {
-        $product = Product::create($request->all());
+        // se verifica si el company_id sea verdaderamente de una empresa
+        $user = User::find($request->get('company_id'));
+        if ($user->isCompany()) {
+            $product = Product::create($request->all());
 
-        // verifico si la empresa del producto recien creado esta asignada a 
-        // algún vendedor buscando un producto que no sea el que se acaba de 
-        // guardar y verificando si tiene vendedores
-        $other_product = $product->company->products()->where('id', '!=', $product->id)
-                ->where('status_id', 1)->first();
-
-        if ($other_product->sellers()->exists()) {
-            // si existen vendedores, se guardan sus id 
-            $sellers_id = $other_product->sellers()->lists('user_id');
-            // como sellers_id no es un array , ya que lists() devuelve un 
-            // array dentro de un objeto, se deben pasar los id a otro array
-            foreach ($sellers_id as $id) {
-                $arrayId [] = $id; 
+            // se guarda otro producto distinto al que se creó, si no hay guarda null
+            $other_product = $product->company->products()
+                    ->where('id', '!=', $product->id)
+                    ->where('status_id', 1)->first();
+            // se verifica si other_product es nulo o no
+            if (!is_null($other_product)) {
+                // si es mo es nulo verifico si la empresa del producto recien  
+                // creado esta asignada a algún vendedor, verificando 
+                // si other_product tiene vendedores
+                if ($other_product->sellers()->exists()) {
+                    // si existen vendedores, se guardan sus id 
+                    $sellers_id = $other_product->sellers()->lists('user_id');
+                    // como sellers_id no es un array , ya que lists() devuelve un 
+                    // array dentro de un objeto, se deben pasar los id a otro array
+                    foreach ($sellers_id as $id) {
+                        $arrayId [] = $id; 
+                    }
+                    // se agrega a la tabla pivote (product_user) un registro con el id 
+                    // de cada vendedor, y el id del producto recien creado
+                    $product->sellers()->attach($arrayId);
+                }
             }
-            // se agrega a la tabla pivote (product_user) un registro con el id 
-            // de cada vendedor, y el id del producto recien creado
-            $product->sellers()->attach($arrayId);
+            $message = $product->name.trans('messages.new');
+            Session::flash('new', $message);
+            return redirect()->route('dashboard.products.index');
         }
-
-        $message = $product->name.trans('messages.new');
-      
-        Session::flash('new', $message);
-    
-        return redirect()->route('dashboard.products.index');
+        Session::flash('error', trans('messages.error'));
+        return redirect()->back();
     }
 
     /**
@@ -154,48 +161,50 @@ class ProductsController extends Controller
     {
         $product = Product::findOrFail($id);
 
-        // se verifica si va haber cambio de empresa
-        if ($request->get('company_id') != $product->company_id) {
-            // si la empresa que se va a guardar es distinta a la que está guardada
-            // se busca la nueva empresa
-            $new_company = User::findOrFail($request->get('company_id'));
-            // se busca el primer producto de la empresa que esté activo
-            $new_company_product = $new_company->products()->where('status_id', 1)->first();
-            // se verifica si ese producto tiene vendedores
-            if ($new_company_product->sellers()->exists()) {
-                // si existen vendedores para $new_company_product, se guardan
-                $sellers = $new_company_product->sellers()->get();
-                // se colocan los id de dichos vendedores en arrayId
-                foreach ($sellers as $seller) {
-                    $arrayId [] = $seller->id; 
+        // se verifica si el company_id sea verdaderamente de una empresa
+        $company = User::find($request->get('company_id'));
+        if ($company->isCompany()) {
+            // se verifica si va haber cambio de empresa
+            if ($request->get('company_id') != $product->company_id) {
+                // si la empresa que se va a guardar es distinta a la que está guardada
+                // $company se convierte en la nueva empresa
+                // se busca el primer producto de la empresa que esté activo
+                $new_company_product = $company->products()->where('status_id', 1)->first();
+                // se verifica que $new_company_product no sea nulo
+                if (!is_null($new_company_product)) {
+                    // si no es nulo se verifica si ese producto tiene vendedores
+                    if ($new_company_product->sellers()->exists()) {
+                        // si existen vendedores para $new_company_product, se guardan
+                        $sellers = $new_company_product->sellers()->get();
+                        // se colocan los id de dichos vendedores en arrayId
+                        foreach ($sellers as $seller) {
+                            $arrayId [] = $seller->id; 
+                        }
+                        // se sincronizan los vendedores que posee la nueva empresa
+                        $product->sellers()->sync($arrayId);
+                    }
+                    else {
+                        // si new_company_product no tiene vendedores, se 
+                        // eliminan los vinculos de los vendedores al producto
+                        // que se está editando, tenga o no tenga vinculados
+                        $product->sellers()->detach();
+                    }
                 }
-                // se sincronizan los vendedores que posee la nueva empresa
-                $product->sellers()->sync($arrayId);
-            }
-            else {
-                $product->sellers()->detach(); // probar asi solo
-                // si no tiene vendedores, se verifica si el producto que se 
-                // quiere actualizar posee vendedores, es decir, si se encuentra 
-                // en la tabla pivote
-                /*if ($product->sellers()->exists()) {
-                    // si es asi, se debe quitar de la tabla pivote ya que 
-                    // $new_company_product no pertenece y por lo tanto $product tampoco 
-                    // deberia ya que la empresa a la que pertenece no esta 
-                    // asignada a ningun vendedor, se eliminan todos los vendedores de $product
+                else {
+                    // si es nulo, quiere decir que la nueva empresa no tiene productos
+                    // o tiene productos pero no estan activos. Por lo tanto al 
+                    // producto que se está editando se le eliminan todos los 
+                    // vendedores vinculados, tenga o no tenga
                     $product->sellers()->detach();
-                    // $product no estará vinculado a ningún vendedor
-                }*/
+                }
             }
+            $product->fill($request->all());
+            $product->save();
+            $message = $product->name.trans('messages.edit');
+            Session::flash('edit', $message);
+            return redirect()->back();
         }
-
-        $product->fill($request->all());
-
-        $product->save();
-
-        $message = $product->name.trans('messages.edit');
-
-        Session::flash('edit', $message);
-
+        Session::flash('error', trans('messages.error'));
         return redirect()->back();
     }
